Create and view customers
This involved some work around how data is stored in vuex, firstly to
simplify it by storing the response object directly instead of mapping
it and secondly to allow dynamically creating new modules so a new
module can be created per customer. Initially I attempted to use one
module for all customers, with an array of customers, but the getters
had to return a function so you could look up the values for a
customer based on the ID, and this wasn't reactive.

Unfortunately vuex doesn't have a way of checking if a module is already
loaded which is required to stop both error messages, and the stored
data being cleared when you want to view the data again. This means I
have to inspect the internals of vuex to see if the module should be
created or not which is brittle but the only way to have a reactive view
and to use existing data instead of waiting for an API request to load a
page if the data is already available locally.