Generate resource IDs client-side
The server has been changed to handle non-unique IDs in a way so as to
have no effect on other tenants (other tailoring companies). So now we
can generate IDs on the client side without fear of security problems
and we can now use these IDs for connecting up all data even if the data
doesn't yet exist on the server, allowing for offline usage.

Offline usage hasn't been set up yet, but with this change it's now
possible to do so.