Update company's unit of measurement.
This required splitting the user/company in to separate stores so the
company stored in the company store can be updated and PUT to the API
separately from the user.

As a result of the splitting it became necessary to add more cohesion to
the way authentication is handled because of increasing complexity so
this has also been moved to using the beforeEnter handler on the router,
meaning the page won't load until the authentication handler has
completed and stored a user in the store if the page requires it.